Refactor EmulatorPatchForm to use new IntegratedPatchForm for reusability
